+++ src/mjlab/asset_zoo/robots/limx_hud03/__init__.py
@@ -1,0 +1,1 @@
+"""LimX Dynamics HUD03 humanoid."""